refactor(dogstatsd): migrate status component to v2
