+++ comp/dogstatsd/bundle.go
@@ -10,7 +10,7 @@ import (
 	replayfx "github.com/DataDog/datadog-agent/comp/dogstatsd/replay/fx"
 	"github.com/DataDog/datadog-agent/comp/dogstatsd/server"
 	"github.com/DataDog/datadog-agent/comp/dogstatsd/serverDebug/serverdebugimpl"
-	"github.com/DataDog/datadog-agent/comp/dogstatsd/status/statusimpl"
+	statusfx "github.com/DataDog/datadog-agent/comp/dogstatsd/status/fx"
 	offlinereporterfx "github.com/DataDog/datadog-agent/comp/offlinereporter/fx"
 	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
 )
@@ -25,5 +25,5 @@ func Bundle(params server.Params) fxutil.BundleOptions {
 		replayfx.Module(),
 		pidmapimpl.Module(),
 		server.Module(params),
-		statusimpl.Module())
+		statusfx.Module())
 }
