--- comp/dogstatsd/status/BUILD.bazel
@@ -1,8 +1,0 @@
-load("@rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "status",
-    srcs = ["component.go"],
-    importpath = "github.com/DataDog/datadog-agent/comp/dogstatsd/status",
-    visibility = ["//visibility:public"],
-)
