+++ comp/dogstatsd/status/def/component.go
