+++ comp/dogstatsd/status/impl/status_test.go
@@ -22,19 +22,19 @@ func TestStatusDisabledWhenADPEnabled(t *testing.T) {
 	config.Set("data_plane.enabled", true, configmodel.SourceAgentRuntime)
 	config.Set("data_plane.dogstatsd.enabled", true, configmodel.SourceAgentRuntime)
 
-	deps := dependencies{
+	deps := Requires{
 		Config: config,
 	}
-	provides := newStatusProvider(deps)
+	provides := NewComponent(deps)
 
 	assert.Nil(t, provides.Status.Provider)
 }
 
 func TestStatusOutputPresent(t *testing.T) {
-	deps := dependencies{
+	deps := Requires{
 		Config: configmock.New(t),
 	}
-	provides := newStatusProvider(deps)
+	provides := NewComponent(deps)
 
 	statusProvider := provides.Status.Provider
